refactor(#30): move /repositories/interfaces to /repositories
