+++ backend/zerowaste/src/main/java/com/zerowaste/zerowaste/repositories/UsersRepository.java
@@ -1,4 +1,4 @@
-package com.zerowaste.zerowaste.repositories.interfaces;
+package com.zerowaste.zerowaste.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
