+++ backend/zerowaste/src/main/java/com/zerowaste/zerowaste/services/users/AuthenticateUserService.java
@@ -26,7 +26,7 @@ import com.auth0.jwt.exceptions.JWTVerificationException;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.zerowaste.zerowaste.models.user.User;
-import com.zerowaste.zerowaste.repositories.interfaces.UsersRepository;
+import com.zerowaste.zerowaste.repositories.UsersRepository;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
